requirements.txt
tensorflow
pillow
numpy
fastapi
aiofiles
python-multipart
uvicorn
